src/requirements-rs5002mqtt.txt
hidapi==0.9.0.post2
paho-mqtt>=1.6.0
